Few usages with users
Now each user can enter or not to some properties of products such as deleting it. This is only implemented in WEB, the API REST IS NOT IMPLEMENTATED.

We need to implement if a user is not logged he only can see the comments. If is logged and the comment is from him, he can delete it. If is the ADMIN can delete any comment.

Also, we need to include a way to create new users (now the are only 2, "manolo" with ROLE_USER and "lucia" with ROLE_ADMIN).